# task-api/api/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.0.1",
    "title" : "Sample Api"
  },
  "host" : "localhost:10010",
  "basePath" : "/api/v1",
  "schemes" : [ "http", "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/squares" : {
      "get" : {
        "summary" : "Get Squares",
        "description" : "Provide array of squares id",
        "operationId" : "getSquares",
        "parameters" : [ {
          "name" : "zoom",
          "in" : "query",
          "description" : "The value that determines the scale of approximation.",
          "required" : false,
          "type" : "number",
          "format" : "integer"
        }, {
          "name" : "latitude",
          "in" : "query",
          "description" : "Coordinate determining the position of the point on the surface of the Earth relative to the equator in degrees.",
          "required" : false,
          "type" : "number",
          "format" : "double"
        }, {
          "name" : "longitude",
          "in" : "query",
          "description" : "Coordinate determining the position of the point on the surface of the Earth relative to the zero meridian in degrees.",
          "required" : false,
          "type" : "number",
          "format" : "double"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/squareId"
              }
            }
          },
          "500" : {
            "description" : "An unexpected error occured.",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "x-swagger-router-controller" : "squaresController"
    },
    "/squares/{squareId}" : {
      "get" : {
        "summary" : "Gets array of objects sights",
        "description" : "array of sights in current square",
        "operationId" : "getMapSights",
        "parameters" : [ {
          "name" : "squareId",
          "in" : "path",
          "description" : "Unique id of the sight.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "array of sights in current square",
            "schema" : {
              "$ref" : "#/definitions/mapSights"
            }
          },
          "404" : {
            "description" : "The Sight does not exist"
          }
        }
      },
      "x-swagger-router-controller" : "squaresController"
    },
    "/sights" : {
      "get" : {
        "summary" : "Gets array of sorted sights",
        "description" : "array of sorted sights by poopularity",
        "operationId" : "getMediaSights",
        "parameters" : [ {
          "name" : "sightIds",
          "in" : "query",
          "description" : "string of sightsids",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "page number",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "authToken",
          "in" : "query",
          "description" : "auth token",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "array",
            "schema" : {
              "$ref" : "#/definitions/mediaSights"
            }
          },
          "404" : {
            "description" : "The Sight does not exist"
          }
        }
      },
      "x-swagger-router-controller" : "mediaSightsController"
    },
    "/sights/{sightId}" : {
      "get" : {
        "summary" : "Gets array of objects sights",
        "description" : "array of sights in current square",
        "operationId" : "getMediaSightDescription",
        "parameters" : [ {
          "name" : "sightId",
          "in" : "path",
          "description" : "sightId",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "authToken",
          "in" : "query",
          "description" : "auth token",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "full description of sight",
            "schema" : {
              "$ref" : "#/definitions/sightDescription"
            }
          },
          "404" : {
            "description" : "The Sight does not exist"
          }
        }
      },
      "x-swagger-router-controller" : "mediaSightsController"
    },
    "/swagger" : {
      "x-swagger-pipe" : "swagger_raw"
    }
  },
  "definitions" : {
    "squareId" : {
      "type" : "string",
      "description" : "squareId"
    },
    "mapSight" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique id of a sight."
        },
        "name" : {
          "type" : "string",
          "description" : "Full name of sight."
        },
        "photo" : {
          "type" : "string",
          "description" : "Link to image sights."
        },
        "latitude" : {
          "type" : "number",
          "format" : "double",
          "description" : "Latitude number"
        },
        "longitude" : {
          "type" : "number",
          "format" : "double",
          "description" : "Longitude number"
        }
      }
    },
    "mapSights" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/mapSight"
      }
    },
    "mediaSight" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique id of a sight."
        },
        "name" : {
          "type" : "string",
          "description" : "Full name of sight."
        },
        "photo" : {
          "type" : "string",
          "description" : "Link to image sights."
        },
        "popularity" : {
          "type" : "number",
          "format" : "integer",
          "description" : "popularity number"
        }
      }
    },
    "mediaSights" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/mediaSight"
      }
    },
    "sightDescription" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique id of a sight."
        },
        "description" : {
          "type" : "string",
          "description" : "description"
        }
      }
    },
    "Error" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  },
  "parameters" : {
    "sightId" : {
      "name" : "sightId",
      "in" : "path",
      "description" : "sightId",
      "required" : true,
      "type" : "string"
    },
    "sightIds" : {
      "name" : "sightIds",
      "in" : "query",
      "description" : "string of sightsids",
      "required" : false,
      "type" : "string"
    },
    "page" : {
      "name" : "page",
      "in" : "query",
      "description" : "page number",
      "required" : false,
      "type" : "string"
    },
    "authToken" : {
      "name" : "authToken",
      "in" : "query",
      "description" : "auth token",
      "required" : false,
      "type" : "string"
    },
    "zoom" : {
      "name" : "zoom",
      "in" : "query",
      "description" : "The value that determines the scale of approximation.",
      "required" : false,
      "type" : "number",
      "format" : "integer"
    },
    "latitude" : {
      "name" : "latitude",
      "in" : "query",
      "description" : "Coordinate determining the position of the point on the surface of the Earth relative to the equator in degrees.",
      "required" : false,
      "type" : "number",
      "format" : "double"
    },
    "longitude" : {
      "name" : "longitude",
      "in" : "query",
      "description" : "Coordinate determining the position of the point on the surface of the Earth relative to the zero meridian in degrees.",
      "required" : false,
      "type" : "number",
      "format" : "double"
    },
    "squareId" : {
      "name" : "squareId",
      "in" : "path",
      "description" : "Unique id of the sight.",
      "required" : true,
      "type" : "string"
    }
  },
  "responses" : {
    "Standard500ErrorResponse" : {
      "description" : "An unexpected error occured.",
      "schema" : {
        "$ref" : "#/definitions/Error"
      }
    }
  }
}